fix rebase errors
