+++ qiskit_experiments/library/quantum_volume/qv_experiment.py
@@ -14,7 +14,8 @@ Quantum Volume Experiment class.
 """
 
 from typing import Union, Sequence, Optional, List
-from numpy.random import Generator, default_rng, BitGenerator, SeedSequence
+from numpy.random import Generator, default_rng
+from numpy.random.bit_generator import BitGenerator, SeedSequence
 
 try:
     from qiskit import Aer
