+++ qiskit_experiments/library/randomized_benchmarking/clifford_utils.py
@@ -676,293 +676,6 @@ def _layer_indices_from_num(num: Integral) -> Tuple[Integral, Integral, Integral
     return idx0, idx1, idx2
 
 
-# Constant mapping from 1Q single Clifford gate to 1Q Clifford numerical identifier.
-# This table must be generated using `data.generate_clifford_data.gen_cliff_single_1q_gate_map`, or,
-# equivalently, correspond to the ordering implicitly defined by CliffUtils.clifford_1_qubit_circuit.
-_CLIFF_SINGLE_GATE_MAP_1Q = {
-    ("id", (0,)): 0,
-    ("h", (0,)): 1,
-    ("sxdg", (0,)): 2,
-    ("s", (0,)): 4,
-    ("x", (0,)): 6,
-    ("sx", (0,)): 8,
-    ("y", (0,)): 12,
-    ("z", (0,)): 18,
-    ("sdg", (0,)): 22,
-}
-# Constant mapping from 2Q single Clifford gate to 2Q Clifford numerical identifier.
-# This table must be generated using `data.generate_clifford_data.gen_cliff_single_2q_gate_map`, or,
-# equivalently, correspond to the ordering defined by _layer_indices_from_num and _CLIFFORD_LAYER.
-_CLIFF_SINGLE_GATE_MAP_2Q = {
-    ("id", (0,)): 0,
-    ("id", (1,)): 0,
-    ("h", (0,)): 5760,
-    ("h", (1,)): 2880,
-    ("sxdg", (0,)): 6720,
-    ("sxdg", (1,)): 3200,
-    ("s", (0,)): 7680,
-    ("s", (1,)): 3520,
-    ("x", (0,)): 4,
-    ("x", (1,)): 1,
-    ("sx", (0,)): 6724,
-    ("sx", (1,)): 3201,
-    ("y", (0,)): 8,
-    ("y", (1,)): 2,
-    ("z", (0,)): 12,
-    ("z", (1,)): 3,
-    ("sdg", (0,)): 7692,
-    ("sdg", (1,)): 3523,
-    ("cx", (0, 1)): 16,
-    ("cx", (1, 0)): 2336,
-    ("cz", (0, 1)): 368,
-    ("cz", (1, 0)): 368,
-}
-
-
-########
-# Functions for 1-qubit integer Clifford operations
-def compose_1q(lhs: Integral, rhs: Integral) -> Integral:
-    """Return the composition of 1-qubit clifford integers."""
-    return _CLIFFORD_COMPOSE_1Q[lhs, rhs]
-
-
-def inverse_1q(num: Integral) -> Integral:
-    """Return the inverse of a 1-qubit clifford integer."""
-    return _CLIFFORD_INVERSE_1Q[num]
-
-
-def num_from_1q_circuit(qc: QuantumCircuit) -> Integral:
-    """Convert a given 1-qubit Clifford circuit to the corresponding integer."""
-    num = 0
-    for inst in qc:
-        rhs = _num_from_1q_gate(op=inst.operation)
-        num = _CLIFFORD_COMPOSE_1Q[num, rhs]
-    return num
-
-
-def _num_from_1q_gate(op: Instruction) -> int:
-    """
-    Convert a given 1-qubit clifford operation to the corresponding integer.
-    Note that supported operations are limited to ones in :const:`CLIFF_SINGLE_GATE_MAP_1Q` or Rz gate.
-
-    Args:
-        op: operation to be converted.
-
-    Returns:
-        An integer representing a Clifford consisting of a single operation.
-
-    Raises:
-        QiskitError: if the input instruction is not a Clifford instruction.
-        QiskitError: if rz is given with a angle that is not Clifford.
-    """
-    if op.name in {"delay", "barrier"}:
-        return 0
-    try:
-        name = _deparameterized_name(op)
-        return _CLIFF_SINGLE_GATE_MAP_1Q[(name, (0,))]
-    except QiskitError as err:
-        raise QiskitError(
-            f"Parameterized instruction {op.name} could not be converted to integer Clifford"
-        ) from err
-    except KeyError as err:
-        raise QiskitError(
-            f"Instruction {op.name} could not be converted to integer Clifford"
-        ) from err
-
-
-def _deparameterized_name(inst: Instruction) -> str:
-    if inst.name == "rz":
-        if np.isclose(inst.params[0], np.pi) or np.isclose(inst.params[0], -np.pi):
-            return "z"
-        elif np.isclose(inst.params[0], np.pi / 2):
-            return "s"
-        elif np.isclose(inst.params[0], -np.pi / 2):
-            return "sdg"
-        else:
-            raise QiskitError("Wrong param {} for rz in clifford".format(inst.params[0]))
-
-    return inst.name
-
-
-########
-# Functions for 2-qubit integer Clifford operations
-def compose_2q(lhs: Integral, rhs: Integral) -> Integral:
-    """Return the composition of 2-qubit clifford integers."""
-    num = lhs
-    for layer, idx in enumerate(_layer_indices_from_num(rhs)):
-        circ = _CLIFFORD_LAYER[layer][idx]
-        num = _compose_num_with_circuit_2q(num, circ)
-    return num
-
-
-def inverse_2q(num: Integral) -> Integral:
-    """Return the inverse of a 2-qubit clifford integer."""
-    return _CLIFFORD_INVERSE_2Q[num]
-
-
-def num_from_2q_circuit(qc: QuantumCircuit) -> Integral:
-    """Convert a given 2-qubit Clifford circuit to the corresponding integer."""
-    return _compose_num_with_circuit_2q(0, qc)
-
-
-def _compose_num_with_circuit_2q(num: Integral, qc: QuantumCircuit) -> Integral:
-    """Compose a number that represents a Clifford, with a Clifford circuit, and return the
-    number that represents the resulting Clifford."""
-    lhs = num
-    for inst in qc:
-        qubits = tuple(qc.find_bit(q).index for q in inst.qubits)
-        rhs = _num_from_2q_gate(op=inst.operation, qubits=qubits)
-        lhs = _CLIFFORD_COMPOSE_2Q[lhs, rhs]
-    return lhs
-
-
-def _num_from_2q_gate(
-    op: Instruction, qubits: Optional[Union[Tuple[int, int], Tuple[int]]] = None
-) -> int:
-    """
-    Convert a given 1-qubit clifford operation to the corresponding integer.
-    Note that supported operations are limited to ones in `CLIFF_SINGLE_GATE_MAP_2Q` or Rz gate.
-
-    Args:
-        op: operation of instruction to be converted.
-        qubits: qubits to which the operation applies
-
-    Returns:
-        An integer representing a Clifford consisting of a single operation.
-
-    Raises:
-        QiskitError: if the input instruction is not a Clifford instruction.
-        QiskitError: if rz is given with a angle that is not Clifford.
-    """
-    if op.name in {"delay", "barrier"}:
-        return 0
-
-    qubits = qubits or (0, 1)
-    try:
-        name = _deparameterized_name(op)
-        return _CLIFF_SINGLE_GATE_MAP_2Q[(name, qubits)]
-    except QiskitError as err:
-        raise QiskitError(
-            f"Parameterized instruction {op.name} could not be converted to integer Clifford"
-        ) from err
-    except KeyError as err:
-        raise QiskitError(
-            f"Instruction {op.name} on {qubits} could not be converted to integer Clifford"
-        ) from err
-
-
-def _append_v_w(qc, vw0, vw1):
-    if vw0 == "v":
-        qc.sdg(0)
-        qc.h(0)
-    elif vw0 == "w":
-        qc.h(0)
-        qc.s(0)
-    if vw1 == "v":
-        qc.sdg(1)
-        qc.h(1)
-    elif vw1 == "w":
-        qc.h(1)
-        qc.s(1)
-
-
-def _create_cliff_2q_layer_0():
-    """Layer 0 consists of 0 or 1 H gates on each qubit, followed by 0/1/2 V gates on each qubit.
-    Number of Cliffords == 36."""
-    circuits = []
-    num_h = [0, 1]
-    v_w_gates = ["i", "v", "w"]
-    for h0, h1, v0, v1 in itertools.product(num_h, num_h, v_w_gates, v_w_gates):
-        qc = QuantumCircuit(2)
-        for _ in range(h0):
-            qc.h(0)
-        for _ in range(h1):
-            qc.h(1)
-        _append_v_w(qc, v0, v1)
-        circuits.append(qc)
-    return circuits
-
-
-def _create_cliff_2q_layer_1():
-    """Layer 1 consists of one of the following:
-    - nothing
-    - cx(0,1) followed by 0/1/2 V gates on each qubit
-    - cx(0,1), cx(1,0) followed by 0/1/2 V gates on each qubit
-    - cx(0,1), cx(1,0), cx(0,1)
-    Number of Cliffords == 20."""
-    circuits = [QuantumCircuit(2)]  # identity at the beginning
-
-    v_w_gates = ["i", "v", "w"]
-    for v0, v1 in itertools.product(v_w_gates, v_w_gates):
-        qc = QuantumCircuit(2)
-        qc.cx(0, 1)
-        _append_v_w(qc, v0, v1)
-        circuits.append(qc)
-
-    for v0, v1 in itertools.product(v_w_gates, v_w_gates):
-        qc = QuantumCircuit(2)
-        qc.cx(0, 1)
-        qc.cx(1, 0)
-        _append_v_w(qc, v0, v1)
-        circuits.append(qc)
-
-    qc = QuantumCircuit(2)  # swap at the end
-    qc.cx(0, 1)
-    qc.cx(1, 0)
-    qc.cx(0, 1)
-    circuits.append(qc)
-    return circuits
-
-
-def _create_cliff_2q_layer_2():
-    """Layer 2 consists of a Pauli gate on each qubit {Id, X, Y, Z}.
-    Number of Cliffords == 16."""
-    circuits = []
-    pauli = ("i", XGate(), YGate(), ZGate())
-    for p0, p1 in itertools.product(pauli, pauli):
-        qc = QuantumCircuit(2)
-        if p0 != "i":
-            qc.append(p0, [0])
-        if p1 != "i":
-            qc.append(p1, [1])
-        circuits.append(qc)
-    return circuits
-
-
-_CLIFFORD_LAYER = (
-    _create_cliff_2q_layer_0(),
-    _create_cliff_2q_layer_1(),
-    _create_cliff_2q_layer_2(),
-)
-_NUM_LAYER_0 = 36
-_NUM_LAYER_1 = 20
-_NUM_LAYER_2 = 16
-
-
-@lru_cache(maxsize=None)
-def _transformed_clifford_layer(
-    layer: int, index: Integral, basis_gates: Tuple[str, ...]
-) -> QuantumCircuit:
-    # Return the index-th quantum circuit of the layer translated with the basis_gates.
-    # The result is cached for speed.
-    return _synthesize_clifford_circuit(_CLIFFORD_LAYER[layer][index], basis_gates)
-
-
-def _num_from_layer_indices(triplet: Tuple[Integral, Integral, Integral]) -> Integral:
-    """Return the clifford number corresponding to the input triplet."""
-    num = triplet[0] * _NUM_LAYER_1 * _NUM_LAYER_2 + triplet[1] * _NUM_LAYER_2 + triplet[2]
-    return num
-
-
-def _layer_indices_from_num(num: Integral) -> Tuple[Integral, Integral, Integral]:
-    """Return the triplet of layer indices corresponding to the input number."""
-    idx2 = num % _NUM_LAYER_2
-    num = num // _NUM_LAYER_2
-    idx1 = num % _NUM_LAYER_1
-    idx0 = num // _NUM_LAYER_1
-    return idx0, idx1, idx2
-
-
 def compute_target_bitstring(self, circuit: QuantumCircuit) -> str:
     """For a Clifford circuit C, compute C|0>.
     Args:
