+++ qiskit_experiments/library/randomized_benchmarking/rb_experiment.py
@@ -19,7 +19,8 @@ from numbers import Integral
 from typing import Union, Iterable, Optional, List, Sequence, Tuple
 
 import numpy as np
-from numpy.random import Generator, default_rng, BitGenerator, SeedSequence
+from numpy.random import Generator, default_rng
+from numpy.random.bit_generator import BitGenerator, SeedSequence
 
 from qiskit.circuit import QuantumCircuit, Instruction, Barrier
 from qiskit.exceptions import QiskitError
