--- releasenotes/notes/fix-992-1d3ab6862e7578ac.yaml
@@ -1,8 +1,0 @@
----
-fixes:
-  - |
-    Fixes bug in the :class:`~.LocalReadoutError` experiment where analysis
-    would fail when run on an ideal simulator with no readout error.
-
-    See Issue #992 <https://github.com/Qiskit/qiskit-experiments/issues/992>`_
-    for additional details.
